wip: split performance store schema helpers
